+++ lib/braintree/connected_merchant_paypal_status_changed.rb
@@ -6,6 +6,8 @@ module Braintree
     attr_reader :merchant_public_id
     attr_reader :oauth_application_client_id
 
+    alias_method :merchant_id, :merchant_public_id
+
     def initialize(attributes)
       set_instance_variables_from_hash(attributes)
     end