+++ lib/braintree/transaction_gateway.rb
@@ -215,11 +215,12 @@ module Braintree
           :add_billing_address_to_payment_method,
           :store_shipping_address_in_vault,
           :venmo_sdk_session,
+          :payee_id,
           :payee_email,
           :skip_advanced_fraud_checking,
           :skip_avs,
           :skip_cvv,
-          {:paypal => [:custom_field, :payee_email, :description, {:supplementary_data => :_any_key_}]},
+          {:paypal => [:custom_field, :payee_id, :payee_email, :description, {:supplementary_data => :_any_key_}]},
           {:three_d_secure => [:required]},
           {:amex_rewards => [:request_id, :points, :currency_amount, :currency_iso_code]},
           {:venmo => [:profile_id]}
@@ -227,7 +228,7 @@ module Braintree
         },
         {:custom_fields => :_any_key_},
         {:descriptor => [:name, :phone, :url]},
-        {:paypal_account => [:email, :token, :paypal_data, :payee_email]},
+        {:paypal_account => [:email, :token, :paypal_data, :payee_id, :payee_email]},
         {:industry => [:industry_type, {:data => [:folio_number, :check_in_date, :check_out_date, :travel_package, :lodging_check_in_date, :lodging_check_out_date, :departure_date, :lodging_name, :room_rate]}]},
         {:apple_pay_card => [:number, :cardholder_name, :cryptogram, :expiration_month, :expiration_year, :eci_indicator]},
         {:android_pay_card => [:number, :cryptogram, :google_transaction_id, :expiration_month, :expiration_year, :source_card_type, :source_card_last_four, :eci_indicator]}