+++ spec/integration/braintree/transaction_spec.rb
@@ -1046,6 +1046,20 @@ describe Braintree::Transaction do
     end
 
     context "transaction_source" do
+      it "marks a transactions as recurring_first" do
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => Braintree::Test::TransactionAmounts::Authorize,
+          :credit_card => {
+            :number => Braintree::Test::CreditCardNumbers::Visa,
+            :expiration_date => "12/12",
+          },
+          :transaction_source => "recurring_first"
+        )
+        result.success?.should == true
+        result.transaction.recurring.should == true
+      end
+
       it "marks a transactions as recurring" do
         result = Braintree::Transaction.create(
           :type => "sale",
@@ -1060,6 +1074,20 @@ describe Braintree::Transaction do
         result.transaction.recurring.should == true
       end
 
+      it "marks a transactions as merchant" do
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => Braintree::Test::TransactionAmounts::Authorize,
+          :credit_card => {
+            :number => Braintree::Test::CreditCardNumbers::Visa,
+            :expiration_date => "12/12",
+          },
+          :transaction_source => "merchant"
+        )
+        result.success?.should == true
+        result.transaction.recurring.should == false
+      end
+
       it "marks a transactions as moto" do
         result = Braintree::Transaction.create(
           :type => "sale",
@@ -1072,6 +1100,20 @@ describe Braintree::Transaction do
         )
         result.success?.should == true
         result.transaction.recurring.should == false
+      end
+
+      it "handles validation when transaction source invalid" do
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => Braintree::Test::TransactionAmounts::Authorize,
+          :credit_card => {
+            :number => Braintree::Test::CreditCardNumbers::Visa,
+            :expiration_date => "12/12",
+          },
+          :transaction_source => "invalid_value"
+        )
+        result.success?.should == false
+        result.errors.for(:transaction).on(:transaction_source)[0].code.should == Braintree::ErrorCodes::Transaction::TransactionSourceIsInvalid
       end
     end
 
@@ -1542,6 +1584,81 @@ describe Braintree::Transaction do
         )
         result.success?.should == true
         result.transaction.should_not be_nil
+      end
+
+      it "can create a transaction with a payee id" do
+        customer = Braintree::Customer.create!
+        nonce = nonce_for_new_payment_method(
+          :paypal_account => {
+            :consent_code => "PAYPAL_CONSENT_CODE",
+          }
+        )
+        nonce.should_not be_nil
+
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => Braintree::Test::TransactionAmounts::Authorize,
+          :payment_method_nonce => nonce,
+          :paypal_account => {
+            :payee_id => "fake-payee-id"
+          }
+        )
+
+        result.success?.should == true
+        result.transaction.paypal_details.should_not be_nil
+        result.transaction.paypal_details.debug_id.should_not be_nil
+        result.transaction.paypal_details.payee_id.should == "fake-payee-id"
+      end
+
+      it "can create a transaction with a payee id in the options params" do
+        customer = Braintree::Customer.create!
+        nonce = nonce_for_new_payment_method(
+          :paypal_account => {
+            :consent_code => "PAYPAL_CONSENT_CODE",
+          }
+        )
+        nonce.should_not be_nil
+
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => Braintree::Test::TransactionAmounts::Authorize,
+          :payment_method_nonce => nonce,
+          :paypal_account => {},
+          :options => {
+            :payee_id => "fake-payee-id"
+          }
+        )
+
+        result.success?.should == true
+        result.transaction.paypal_details.should_not be_nil
+        result.transaction.paypal_details.debug_id.should_not be_nil
+        result.transaction.paypal_details.payee_id.should == "fake-payee-id"
+      end
+
+      it "can create a transaction with a payee id in options.paypal" do
+        customer = Braintree::Customer.create!
+        nonce = nonce_for_new_payment_method(
+          :paypal_account => {
+            :consent_code => "PAYPAL_CONSENT_CODE",
+          }
+        )
+        nonce.should_not be_nil
+
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => Braintree::Test::TransactionAmounts::Authorize,
+          :payment_method_nonce => nonce,
+          :options => {
+            :paypal => {
+              :payee_id => "fake-payee-id"
+            }
+          }
+        )
+
+        result.success?.should == true
+        result.transaction.paypal_details.should_not be_nil
+        result.transaction.paypal_details.debug_id.should_not be_nil
+        result.transaction.paypal_details.payee_id.should == "fake-payee-id"
       end
 
       it "can create a transaction with a payee email" do