+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.89.0
+* Warn when instantiating a `Braintree::Gateway` with mixed environments
+* Allow payee ID to be passed in options params for transaction create
+* Add `merchant_id` to `ConnectedMerchantStatusTransitioned` and `ConnectedMerchantPayPalStatusChanged` Auth webhooks
+
 # 2.88.0
 * Add support for Micro Transfer ACH verifications
 * Add `image_url` and `token` attributes to `AndroidPayDetails` and `ApplePayDetails`